删除产品url判空

+++ buluerp_vue/ruoyi-admin/src/main/java/com/ruoyi/web/service/impl/ErpProductsServiceImpl.java
@@ -155,8 +155,10 @@ public class ErpProductsServiceImpl extends ServiceImpl<ErpProductsMapper, ErpPr
             erpPackagingListMapper.deleteErpPackagingListByProductId(erpProducts.getId());
             erpProductsMapper.clearProductMaterial(erpProducts.getId());
             String url=erpProducts.getPictureUrl();
-            url=parseActualPath(url);
-            FileUtils.deleteFile(url);
+            if(!StringUtils.isBlank(url)){
+                url=parseActualPath(url);
+                FileUtils.deleteFile(url);
+            }
         }
         return erpProductsMapper.deleteBatchIds(ids);
     }
